examples/pytorch_lightning/requirements.txt
pytorch-lightning==2.0.1
omegaconf
filelock
torchmetrics
torchvision
